refactor: move format_x_info to respective subdir in utils and ensure parity between instance and cluster tools
